Add useLocalStorage and useNearScreen custom hooks Create custom hooks on functions for the store of likes on local storage and to render elements with viewport
