+++ src/components/PhotoCard/index.js
@@ -1,47 +1,15 @@
-import React, { useRef, useEffect, useState } from 'react'
+import React from 'react'
 import { Article, ImageWrapper, Image, Button } from './styles'
 import { HiHeart, HiOutlineHeart } from 'react-icons/hi'
-
+import { useLocalStorage } from '../../hooks/useLocalStorage'
+import { useNearScreen } from '../../hooks/useNearScreen'
 const DEFAULT_IMAGE = 'https://images.unsplash.com/photo-1518791841217-8f162f1e1131?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=800&q=60'
 
 export const PhotoCard = ({ id, likes = 0, src = DEFAULT_IMAGE }) => {
-  const ref = useRef(null)
-  const [show, setShow] = useState(false)
+  const [show, ref] = useNearScreen()
   const key = `like-${id}`
-  const [liked, setLiked] = useState(() => {
-    try {
-      const like = window.localStorage.getItem(key)
-      return JSON.parse(like)
-    } catch (e) {
-      return false
-    }
-  })
-
-  useEffect(function () {
-    Promise.resolve(
-      typeof window.IntersectionObserver !== 'undefined'
-        ? window.IntersectionObserver
-        : import('intersection-observer')
-    ).then(() => {
-      const observer = new window.IntersectionObserver(function (entries) {
-        const { isIntersecting } = entries[0]
-        if (isIntersecting) {
-          setShow(true)
-          observer.disconnect()
-        }
-      })
-      observer.observe(ref.current)
-    })
-  }, [ref])
+  const [liked, setLiked] = useLocalStorage(key, false)
   const Icon = liked ? HiHeart : HiOutlineHeart
-  const setLocalStorage = value => {
-    try {
-      window.localStorage.setItem(key, value)
-      setLiked(value)
-    } catch (e) {
-      console.log(e)
-    }
-  }
   return (
     <Article ref={ref}>
       {
@@ -53,13 +21,12 @@ export const PhotoCard = ({ id, likes = 0, src = DEFAULT_IMAGE }) => {
 
               </ImageWrapper>
             </a>
-            <Button onClick={() => setLocalStorage(!liked)}>
+            <Button onClick={() => setLiked(!liked)}>
               <Icon size='16px' />
               {likes} Likes!
             </Button>
           </>
       }
-
     </Article>
   )
 }
